fix(fresco): keep render node kind parsing under the length gate
